+++ examples/src/test/java/ai/djl/examples/training/TrainBertTest.java
@@ -12,12 +12,14 @@
  */
 package ai.djl.examples.training;
 
+import ai.djl.translate.TranslateException;
+import java.io.IOException;
 import org.testng.annotations.Test;
 
 public class TrainBertTest {
 
     @Test
-    public void testTrainBert() {
+    public void testTrainBert() throws IOException, TranslateException {
         String[] args = new String[] {"-g", "1", "-m", "1", "-e", "1"};
         TrainBertOnCode.runExample(args);
     }
